[jwies/review-refusal-visibility] review: split agent-definition loading out of dispatch.ts for the line cap
The refusal-fallback additions pushed dispatch.ts to 1002 lines, one over the
max-lines cap that drove the roster split in #304. parseAgentFile/loadAgents
and the DispatchFs shape move to dispatch-agents.ts: self-contained, no
dispatch state, and re-exported so callers keep one import surface.

DispatchFs moves with them rather than staying behind, because leaving it in
dispatch.ts made the two modules import each other and the cycle broke
loadAgents at runtime while tsc stayed clean.
